pyproject.toml
[tool.poetry]
name = "genius"
version = "0.1.0"
description = ""
authors = ["mohammadali koutahi <mmliinjas@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
asgiref = "3.8.1"
certifi = "2024.12.14"
charset-normalizer = "3.4.0"
colorthief = "0.2.1"
django = "5.1.4"
django-allauth = "65.3.0"
idna = "3.10"
pillow = "11.0.0"
requests = "2.32.3"
sqlparse = "0.5.3"
urllib3 = "2.2.3"
spotipy = "^2.24.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
